Reorganize module
